requirements.txt
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.5.0
fonttools==4.55.3
fsspec==2024.12.0
gradio_client==1.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
Pillow==9.5.0
pyglet==1.5.31
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
simple-lama-inpainting==0.1.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
termcolor==2.5.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.0
trimesh==4.5.3
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
